Move status() from CommonClient to LocalClient
Not supported for remote.
