+++ svn/common.py
@@ -12,15 +12,6 @@ import svn.common_base
 
 _LOGGER = logging.getLogger(__name__)
 
-_STATUS_ENTRY = \
-    collections.namedtuple(
-        '_STATUS_ENTRY', [
-            'name',
-            'type_raw_name',
-            'type',
-            'revision',
-        ])
-
 _FILE_HUNK_PREFIX = 'Index: '
 
 _HUNK_HEADER_LEFT_PREFIX = '--- '
@@ -293,42 +284,6 @@ class CommonClient(svn.common_base.CommonBase):
         cmd.append('--force') if force else None
 
         self.run_command('export', cmd)
-
-    def status(self, rel_path=None):
-        full_url_or_path = self.__url_or_path
-        if rel_path is not None:
-            full_url_or_path += '/' + rel_path
-
-        raw = self.run_command(
-            'status',
-            ['--xml', full_url_or_path],
-            do_combine=True)
-
-        root = xml.etree.ElementTree.fromstring(raw)
-
-        list_ = root.findall('target/entry')
-        for entry in list_:
-            entry_attr = entry.attrib
-            name = entry_attr['path']
-
-            wcstatus = entry.find('wc-status')
-            wcstatus_attr = wcstatus.attrib
-
-            change_type_raw = wcstatus_attr['item']
-            change_type = svn.constants.STATUS_TYPE_LOOKUP[change_type_raw]
-
-            # This will be absent if the file is "unversioned". It'll be "-1"
-            # if added but not committed.
-            revision = wcstatus_attr.get('revision')
-            if revision is not None:
-                revision = int(revision)
-
-            yield _STATUS_ENTRY(
-                name=name,
-                type_raw_name=change_type_raw,
-                type=change_type,
-                revision=revision
-            )
 
     def list(self, extended=False, rel_path=None):
         full_url_or_path = self.__url_or_path
